Restore db methods to trait
